cgo: Rewrite conversion from C array to Go array
The code is now (a little) clearer regarding the conversion from C
arrays to Go arrays, when calling C functions. A Go slice header is
handled directly so its Data is the actual C array pointer.

This patch allows the library to use godoc and govet. The previous
syntax was parsed as wrong by those tools.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>