Try to define a valid domain in a read-only connection
When testing the read-only connection, the function tried to create
a domain with an empty XML declaration. The operation failed because the
connection is read-only (which is what the test expects), but if somehow
the read-only connection allows the creation of a domain, that operation
would fail as well because the XML declaration was empty. So the test
was not correct.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>